Refactoring: renaming of generic encoding_filter to more precise encodings2regdeps.
