+++ clients/drcachesim/tools/record_filter_launcher.cpp
@@ -121,10 +121,11 @@ static droption_t<uint64_t> op_trim_after_timestamp(
     "Removes all records from the first TRACE_MARKER_TYPE_TIMESTAMP marker with "
     "timestamp larger than the specified value.");
 
-static droption_t<bool> op_encoding_filter_enabled(
-    DROPTION_SCOPE_FRONTEND, "encoding_filter_enabled", false,
+droption_t<bool> op_encodings2regdeps(
+    DROPTION_SCOPE_FRONTEND, "encodings2regdeps", false,
     "Enable converting the encoding of instructions to synthetic ISA DR_ISA_REGDEPS.",
-    "Enable converting the encoding of instructions to synthetic ISA DR_ISA_REGDEPS.");
+    "When present, converts the encoding of instructions from a real ISA to the "
+    "DR_ISA_REGDEPS synthetic ISA.");
 } // namespace
 
 int
@@ -154,7 +155,7 @@ _tmain(int argc, const TCHAR *targv[])
             op_output_dir.get_value(), op_stop_timestamp.get_value(),
             op_cache_filter_size.get_value(), op_remove_trace_types.get_value(),
             op_remove_marker_types.get_value(), op_trim_before_timestamp.get_value(),
-            op_trim_after_timestamp.get_value(), op_encoding_filter_enabled.get_value(),
+            op_trim_after_timestamp.get_value(), op_encodings2regdeps.get_value(),
             op_verbose.get_value()));
     std::vector<record_analysis_tool_t *> tools;
     tools.push_back(record_filter.get());
